Move compaction task status store tests to core module
