+++ java/core/src/test/java/sleeper/core/tracker/compaction/task/CompactionTaskStatusTestData.java
@@ -14,11 +14,9 @@
  * limitations under the License.
  */
 
-package sleeper.compaction.core.task;
+package sleeper.core.tracker.compaction.task;
 
 import sleeper.core.record.process.RecordsProcessedSummary;
-import sleeper.core.tracker.compaction.task.CompactionTaskFinishedStatus;
-import sleeper.core.tracker.compaction.task.CompactionTaskStatus;
 
 import java.time.Duration;
 import java.time.Instant;
