+++ java/core/src/test/java/sleeper/core/tracker/compaction/task/CompactionTaskStatusesBuilderTest.java
@@ -13,15 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.compaction.core.task;
+package sleeper.core.tracker.compaction.task;
 
 import org.junit.jupiter.api.Test;
 
 import sleeper.core.record.process.RecordsProcessed;
 import sleeper.core.record.process.RecordsProcessedSummary;
-import sleeper.core.tracker.compaction.task.CompactionTaskFinishedStatus;
-import sleeper.core.tracker.compaction.task.CompactionTaskStatus;
-import sleeper.core.tracker.compaction.task.CompactionTaskStatusesBuilder;
 
 import java.time.Instant;
 import java.util.List;
